LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/events/messageevent-constructor.https-expected.txt


PASS Default event values 
PASS MessageEventInit dictionary 
PASS Passing null for ports member 
FAIL ports attribute should be a FrozenArray assert_true: ev.ports should return the same object expected true got false
PASS initMessageEvent operation 
FAIL Passing null for ports parameter to initMessageEvent assert_throws: function "function () {
    ev.initMessageEvent("test", true, false, "testData", "testOrigin", "testId", window, null)
  }" did not throw
FAIL All parameters to initMessageEvent should be mandatory assert_equals: MessageEvent.prototype.initMessageEvent.length should be 8 expected 8 but got 0
FAIL Passing ServiceWorker for source member undefined is not an object (evaluating 'navigator.serviceWorker.getRegistration')
